fix: price u64->u32

+++ programs/vizing_app/src/vizing_omni/emitter.rs
@@ -66,7 +66,7 @@ pub struct LaunchParams {
 pub struct Message {
     pub mode: u8,
     pub target_program: [u8; 32],
-    pub execute_gas_limit: u64,
+    pub execute_gas_limit: u32,
     pub max_fee_per_gas: u64,
     #[max_len(1024)]
     pub signature: Vec<u8>,
